refactor: extract pdf export to lib/pdf-export.ts (-177 lines in app.tsx)
