+++ src/app.tsx
@@ -24,7 +24,6 @@ import { TooltipRoot } from "@/components/primitives";
 import { useDebouncedValue, useFileWatcher, usePersistedState, useShortcuts, useSyncScroll } from "@/hooks";
 import { getCurrentWindow } from "@tauri-apps/api/window";
 import { listen } from "@tauri-apps/api/event";
-import { tempDir } from "@tauri-apps/api/path";
 import { openPath } from "@tauri-apps/plugin-opener";
 import {
   basename,
@@ -33,12 +32,14 @@ import {
   createMarkdownFile,
   dirname,
   estimateTokens,
+  exportPreviewToPdf,
   FS_CONFLICT,
   isMarkdownPath,
   joinPath,
   listFolder,
   pathExists,
   moveEntry,
+  PdfExportError,
   pickFolder,
   pickMarkdownFile,
   pickSaveMarkdown,
@@ -60,15 +61,6 @@ type UndoOp =
   | { kind: "rename"; from: string; to: string }
   | { kind: "create-folder"; path: string }
   | { kind: "create-file"; path: string };
-
-function escapeHtml(s: string): string {
-  return s
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#39;");
-}
 
 export function App() {
   const [source, setSource] = useState<string>(DEMO_MARKDOWN);
@@ -148,186 +140,17 @@ export function App() {
   }, [setSidebarOpen]);
 
   const exportToPdf = useCallback(async () => {
-    // workaround: Tauri 2's WKWebView no-ops window.print(). instead, clone
-    // the already-rendered preview article (which has mermaid svgs, katex
-    // html, and shiki spans already painted), wrap in a standalone html
-    // doc with print-friendly styles, write to OS temp dir, open in the
-    // user's default browser. the browser auto-triggers its print dialog
-    // on load — user picks "Save as PDF" → real native flow.
-    if (!source.trim()) {
-      setLoadError({
-        message: "nothing to export. open or write some markdown first.",
-      });
-      return;
-    }
-
-    const article = document.querySelector<HTMLElement>(".mdv-prose");
-    if (!article) {
-      setLoadError({
-        message: "preview isn't rendered yet. try again in a moment.",
-      });
-      return;
-    }
-
-    const fileName = activePath ? basename(activePath) : undefined;
-    const title = fileName ?? "marka.md export";
-
-    // self-contained print stylesheet — light theme regardless of app theme.
-    // tracks the live prose CSS conceptually but stays decoupled so the
-    // exported pdf has a stable "document" look.
-    const printStyles = `
-      *, *::before, *::after { box-sizing: border-box; }
-      :root {
-        --pfg: #1d1d1f;
-        --pmuted: #6e6e73;
-        --paccent: #e2722e;
-        --pborder: rgba(0, 0, 0, 0.08);
-      }
-      html, body {
-        background: #fff;
-        color: var(--pfg);
-        font-family: -apple-system, BlinkMacSystemFont, "SF Pro Text", "Inter", system-ui, sans-serif;
-        font-size: 15px;
-        line-height: 1.65;
-        margin: 0;
-        padding: 0;
-        -webkit-font-smoothing: antialiased;
-        /* preserve colors when saving to pdf (e.g. shiki code-block backgrounds) */
-        -webkit-print-color-adjust: exact;
-        print-color-adjust: exact;
-      }
-      .doc { max-width: 720px; margin: 0 auto; padding: 18mm 16mm 22mm; }
-      h1, h2, h3, h4 {
-        font-weight: 600;
-        letter-spacing: -0.018em;
-        /* keep headings on the same page as the first paragraph that follows */
-        break-after: avoid;
-        page-break-after: avoid;
-      }
-      h1 { font-size: 2.1em; margin: 0 0 0.5em; }
-      h2 { font-size: 1.55em; margin: 2em 0 0.5em; }
-      h3 { font-size: 1.25em; margin: 1.6em 0 0.5em; }
-      h4 { font-size: 1em; margin: 1.4em 0 0.4em; }
-      p, ul, ol, blockquote, pre, table { margin: 0 0 1em; }
-      p { orphans: 3; widows: 3; }
-      a {
-        color: var(--paccent);
-        text-decoration: none;
-        border-bottom: 1px solid color-mix(in srgb, var(--paccent) 30%, transparent);
-      }
-      code {
-        font-family: "SF Mono", "JetBrains Mono", ui-monospace, monospace;
-        font-size: 0.88em;
-        background: rgba(0, 0, 0, 0.045);
-        padding: 1px 6px;
-        border-radius: 4px;
-      }
-      pre {
-        font-family: "SF Mono", "JetBrains Mono", ui-monospace, monospace;
-        font-size: 12.5px;
-        line-height: 1.55;
-        background: #fafafa;
-        border: 1px solid var(--pborder);
-        border-radius: 8px;
-        padding: 14px 16px;
-        overflow-x: auto;
-        /* don't split code blocks across pages — looks bad and breaks tokens */
-        break-inside: avoid;
-        page-break-inside: avoid;
-      }
-      blockquote, table, .mdv-mermaid {
-        break-inside: avoid;
-        page-break-inside: avoid;
-      }
-      img { max-width: 100%; height: auto; border-radius: 6px; break-inside: avoid; }
-      pre code { background: transparent; padding: 0; font-size: inherit; border-radius: 0; }
-      pre.shiki, pre.shiki * { font-family: inherit; }
-      /* PDF is a "document" — force catppuccin-latte palette regardless of
-         the user's active app theme. shiki emits --shiki-latte vars inline on
-         every span (multi-theme mode), so we just pick that variant for print.
-         Without this rule, spans would fall back to inherited body color
-         (looks like plain monospace text — no syntax colors). */
-      .shiki, .shiki span {
-        background-color: transparent !important;
-        color: var(--shiki-latte) !important;
-        font-style: var(--shiki-latte-font-style, inherit) !important;
-        font-weight: var(--shiki-latte-font-weight, inherit) !important;
-        text-decoration: var(--shiki-latte-text-decoration, inherit) !important;
-      }
-      blockquote {
-        border-left: 3px solid var(--paccent);
-        padding-left: 16px;
-        color: var(--pmuted);
-        font-style: italic;
-      }
-      hr { border: none; border-top: 1px solid var(--pborder); margin: 2em 0; }
-      ul, ol { padding-left: 24px; }
-      li { margin: 0.25em 0; }
-      /* task lists in PDF — hide bullet, checkbox is the marker (#21) */
-      .task-list-item { list-style: none; margin-left: -1.4em; }
-      .task-list-item input[type="checkbox"] {
-        margin-right: 0.5em;
-        accent-color: var(--paccent);
-        width: 0.95em;
-        height: 0.95em;
-        vertical-align: -0.1em;
-      }
-      table { border-collapse: collapse; width: 100%; }
-      th, td { border: 1px solid var(--pborder); padding: 8px 12px; text-align: left; vertical-align: top; }
-      th { background: rgba(0, 0, 0, 0.03); font-weight: 600; }
-      /* mermaid renders as inline SVG inside <pre class="mdv-mermaid"> */
-      .mdv-mermaid { background: transparent; border: 0; padding: 0; text-align: center; }
-      .mdv-mermaid svg { max-width: 100%; height: auto; }
-      /* hide in-app affordances that don't belong in print */
-      .mdv-copy, .mdv-codeblock > .mdv-copy { display: none !important; }
-      /* margin: 0 on @page leaves no room for the browser's auto header /
-         footer (title, date, url, page numbers). all visible whitespace is
-         pushed into .doc padding instead so it still looks like a margined
-         document. users can also uncheck "Headers and footers" in the print
-         dialog's "More settings" if their browser still tries to render them. */
-      @page { margin: 0; size: auto; }
-      @media print {
-        body { padding: 0; }
-        .doc { padding: 18mm 16mm 22mm; }
-      }
-    `;
-
-    const html = `<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>${escapeHtml(title)}</title>
-  <style>${printStyles}</style>
-</head>
-<body>
-  <main class="doc">
-    ${article.outerHTML}
-  </main>
-  <script>
-    // auto-trigger the print dialog once layout settles
-    window.addEventListener("load", () => {
-      setTimeout(() => window.print(), 350);
-    });
-  </script>
-</body>
-</html>`;
-
     try {
-      const dir = (await tempDir()).replace(/[\\/]+$/, "");
-      const safeName = (fileName ?? "export")
-        .replace(/\.md$/i, "")
-        .replace(/[^\w.-]+/g, "-")
-        .slice(0, 60) || "export";
-      // stable name → overwrites on each export instead of piling up in /tmp
-      const tempPath = `${dir}/marka-${safeName}.html`;
-      await writeMarkdown(tempPath, html);
-      await openPath(tempPath);
+      await exportPreviewToPdf({ source, activePath });
     } catch (err) {
+      const message = err instanceof PdfExportError
+        ? err.message
+        : "couldn't export to pdf";
       console.error("marka.md: pdf export failed", err);
-      setLoadError({ message: "couldn't export to pdf — try again, or check disk space" });
+      setLoadError({ message });
     }
   }, [source, activePath]);
+
 
   const toggleFullscreen = useCallback(async () => {
     const win = getCurrentWindow();
