+++ src/lib/index.ts
@@ -14,6 +14,7 @@ export { STORAGE_KEYS, type StorageKey } from "./storage";
 export { buildCommands, type Command, type CommandActions } from "./commands";
 export { estimateTokens, formatTokens } from "./bundle";
 export { startWindowDrag } from "./window-drag";
+export { exportPreviewToPdf, PdfExportError } from "./pdf-export";
 export { IS_MAC, IS_WINDOWS, IS_LINUX, displayKey, shortcutLabel } from "./platform";
 export {
   pickFolder,
